Import Dual Energy X-ray Absorptiometry (DXA) data.

DXA data were acquired from 1999-2006.

nhanesDXA(year, suppl = FALSE, destfile = NULL)

Arguments

year
The year of the data to import, where 1999<=year<=2006.

suppl
If TRUE then retrieve the supplemental data (default=FALSE).

destfile
The name of a destination file. If NULL then the data are imported into the R environment but no file is created.

Value

By default the table is returned as a data frame. When downloading to file, the return argument is the integer code from download.file where 0 means success and non-zero indicates failure to download.

Details

Provide destfile in order to write the data to file. If destfile is not provided then the data will be imported into the R environment.

Examples

dxa_b <- nhanesDXA(2001)
#> Processing SAS dataset DXX_B 	 ..
dxa_c_s <- nhanesDXA(2003, suppl=TRUE)
#> Processing SAS dataset DXX_C_S 	 ..
nhanesDXA(1999, destfile="dxx.xpt")
#> [1] 0