Search for matching table names

Returns a list of table names that match a specified pattern.

nhanesSearchTableNames(
  pattern = NULL,
  ystart = NULL,
  ystop = NULL,
  includerdc = FALSE,
  nchar = 100,
  details = FALSE
)

Arguments

pattern
Pattern of table names to match

ystart
Four digit year of first survey included in search, where ystart >= 1999.

ystop
Four digit year of final survey included in search, where ystop >= ystart.

includerdc
If TRUE then RDC only tables are included (default=FALSE).

nchar
Truncates the variable description to a max length of nchar.

details
If TRUE then complete table information from the comprehensive data list is returned (default=FALSE).

Value

A list of table names that match the pattern.

Details

Searches the Doc File field in the NHANES Comprehensive Data List (see https://wwwn.cdc.gov/nchs/nhanes/search/DataPage.aspx) for tables that match a given name pattern. Only a single pattern may be entered.

Examples

nhanesSearchTableNames('BMX')
#>  [1] "BMX_D" "BMX"   "BMX_E" "BMX_C" "BMX_B" "BMX_F" "BMX_H" "BMX_G" "BMX_I"
#> [10] "BMX_J"
nhanesSearchTableNames('HPVS', includerdc=TRUE, details=TRUE)
#>        Years
#> 1  2009-2010
#> 2  2005-2006
#> 3  2007-2008
#> 4  2005-2006
#> 5  2005-2006
#> 6  2005-2006
#> 7  2007-2008
#> 8  2009-2010
#> 9  2011-2012
#> 10 2005-2006
#> 11 2009-2010
#> 12 2011-2012
#> 13 2013-2014
#> 14 2013-2014
#> 15 2015-2016
#> 16 2015-2016
#> 17 2015-2016
#> 18 2017-2018
#>                                                                                          Data.File.Name
#> 1                                          Human Papillomavirus (HPV) - 6, 11, 16 & 18 Antibody - Serum
#> 2                                  Human Papillomavirus (HPV) - 6, 11, 16 & 18 Antibody - Serum - Youth
#> 3                             Human Papillomavirus (HPV) - 6, 11, 16 & 18 Antibody – Serum: 4-plex CLIA
#> 4                             Human Papillomavirus (HPV) - 6, 11, 16 & 18 Antibody – Serum: 4-plex CLIA
#> 5  Human Papillomavirus (HPV) - Multiplexed 6, 11, 16, 18, 31, 22, 45, 52 & 58 Antibody – Serum: 9-plex
#> 6             Human Papillomavirus (HPV) DNA - Vaginal Swab: Digene Hybrid Capture & Roche Linear Array
#> 7             Human Papillomavirus (HPV) DNA - Vaginal Swab: Digene Hybrid Capture & Roche Linear Array
#> 8             Human Papillomavirus (HPV) DNA - Vaginal Swab: Digene Hybrid Capture & Roche Linear Array
#> 9             Human Papillomavirus (HPV) DNA - Vaginal Swab: Digene Hybrid Capture & Roche Linear Array
#> 10            Human Papillomavirus (HPV) DNA - Vaginal Swab: Digene Hybrid Capture & Roche Linear Array
#> 11    Human Papillomavirus (HPV) DNA - Vaginal Swab: Digene Hybrid Capture & Roche Linear Array - Youth
#> 12    Human Papillomavirus (HPV) DNA - Vaginal Swab: Digene Hybrid Capture & Roche Linear Array - Youth
#> 13                      Human Papillomavirus (HPV) DNA - Vaginal Swab: Roche Cobas & Roche Linear Array
#> 14              Human Papillomavirus (HPV) DNA - Vaginal Swab: Roche Cobas & Roche Linear Array - Youth
#> 15                                 Human Papillomavirus (HPV) DNA - Vaginal Swab: Roche Cobas High-Risk
#> 16                                    Human Papillomavirus (HPV) DNA - Vaginal Swab: Roche Linear Array
#> 17                            Human Papillomavirus (HPV) DNA - Vaginal Swab: Roche Linear Array - Youth
#> 18                            Human Papillomavirus (HPV) DNA - Vaginal Swab: Roche Linear Array - Youth
#>        Doc.File                      Data.File        Date.Published
#> 1  HPVSER_F Doc HPVSER_F Data [XPT - 171.6 KB]         November 2013
#> 2  HPVS_D_R Doc                       RDC Only             July 2013
#> 3  HPVSER_E Doc HPVSER_E Data [XPT - 155.7 KB]         November 2013
#> 4  HPVSER_D Doc HPVSER_D Data [XPT - 151.6 KB]             July 2013
#> 5  HPVSRM_D Doc HPVSRM_D Data [XPT - 302.6 KB]          January 2015
#> 6  HPVSWR_D Doc HPVSWR_D Data [XPT - 694.4 KB]         November 2010
#> 7  HPVSWR_E Doc HPVSWR_E Data [XPT - 677.9 KB]           August 2012
#> 8  HPVSWR_F Doc HPVSWR_F Data [XPT - 725.2 KB]           August 2012
#> 9  HPVSWR_G Doc HPVSWR_G Data [XPT - 661.1 KB]            March 2015
#> 10 HPVSWR_D Doc HPVSWR_D Data [XPT - 694.4 KB] Updated November 2018
#> 11 HPVS_F_R Doc                       RDC Only           August 2012
#> 12 HPVS_G_R Doc                       RDC Only            March 2015
#> 13 HPVSWR_H Doc HPVSWR_H Data [XPT - 716.6 KB]         December 2016
#> 14 HPVS_H_R Doc                       RDC Only         December 2016
#> 15 HPVSWC_I Doc  HPVSWC_I Data [XPT - 33.3 KB]         November 2018
#> 16 HPVSWR_I Doc HPVSWR_I Data [XPT - 667.5 KB]         November 2018
#> 17 HPVS_I_R Doc                       RDC Only         November 2018
#> 18 HPVS_J_R Doc                       RDC Only         December 2020