Search for tables that contain a specified variable.

Returns a list of table names that contain the variable

nhanesSearchVarName(
  varname = NULL,
  ystart = NULL,
  ystop = NULL,
  includerdc = FALSE,
  nchar = 100,
  namesonly = TRUE
)

Arguments

varname
Name of variable to match.

ystart
Four digit year of first survey included in search, where ystart >= 1999.

ystop
Four digit year of final survey included in search, where ystop >= ystart.

includerdc
If TRUE then RDC only tables are included in list (default=FALSE).

nchar
Truncates the variable description to a max length of nchar.

namesonly
If TRUE then only the table names are returned (default=TRUE).

Details

The NHANES Comprehensive Variable List is scanned to find all data tables that contain the given variable name. Only a single variable name may be entered, and only exact matches will be found.

Examples

nhanesSearchVarName('BMXLEG')
#>  [1] "BMX_D" "BMX"   "BMX_E" "BMX_C" "BMX_B" "BMX_F" "BMX_H" "BMX_G" "BMX_I"
#> [10] "BMX_J"
nhanesSearchVarName('BMXHEAD', ystart=2003)
#> [1] "BMX_D" "BMX_E" "BMX_C" "BMX_F" "BMX_H" "BMX_G" "BMX_I" "BMX_J"