Displays a list of variables in the specified NHANES table.

Enables quick display of table variables and their definitions.

nhanesTableVars(
  data_group,
  nh_table,
  details = FALSE,
  nchar = 100,
  namesonly = FALSE
)

Arguments

data_group
The type of survey (DEMOGRAPHICS, DIETARY, EXAMINATION, LABORATORY, QUESTIONNAIRE). Abbreviated terms may also be used: (DEMO, DIET, EXAM, LAB, Q).

nh_table
The name of the specific table to retrieve.

details
If TRUE then all columns in the variable description are returned (default=FALSE).

nchar
The number of characters in the Variable Description to print. Values are limited to 0<=nchar<=128. This is used to enhance readability, cause variable descriptions can be very long.

namesonly
If TRUE then only the variable names are returned (default=FALSE).

Value

The names of the tables in the specified survey group

Details

Data are retrieved via web scraping using html wrappers from package rvest. Each data table contains multiple, sometimes more than 100, fields. It is helpful to list the field descriptions to ascertain quickly if a data table is of interest.

Examples

nhanesTableVars('LAB', 'CBC_E')
#>    Variable.Name                      Variable.Description
#> 1        LBDBANO          Basophils number (1000 cells/uL)
#> 2        LBDEONO        Eosinophils number (1000 cells/uL)
#> 3       LBDLYMNO         Lymphocyte number (1000 cells/uL)
#> 4        LBDMONO           Monocyte number (1000 cells/uL)
#> 5        LBDNENO  Segmented neutrophils num (1000 cell/uL)
#> 6       LBXBAPCT                     Basophils percent (%)
#> 7       LBXEOPCT                   Eosinophils percent (%)
#> 8         LBXHCT                            Hematocrit (%)
#> 9         LBXHGB                         Hemoglobin (g/dL)
#> 10      LBXLYPCT                    Lymphocyte percent (%)
#> 11         LBXMC Mean cell hemoglobin concentration (g/dL)
#> 12      LBXMCHSI                 Mean cell hemoglobin (pg)
#> 13      LBXMCVSI                     Mean cell volume (fL)
#> 14      LBXMOPCT                      Monocyte percent (%)
#> 15       LBXMPSI                 Mean platelet volume (fL)
#> 16      LBXNEPCT         Segmented neutrophils percent (%)
#> 17      LBXPLTSI                     Platelet count (%) SI
#> 18      LBXRBCSI                         Red cell count SI
#> 19        LBXRDW           Red cell distribution width (%)
#> 20      LBXWBCSI               White blood cell count:  SI
#> 21          SEQN               Respondent sequence number.
nhanesTableVars('EXAM', 'OHX_E', details=TRUE, nchar=50)
#>    Variable.Name                               Variable.Description
#> 1       OHAEXSTS                    Overall Oral Health Exam Status
#> 2         OHAPOS Was the SP in a recumbent (laying down) position f
#> 3         OHQ410 Do you have an upper removable partial or full den
#> 4         OHQ420             Do you usually wear it during the day?
#> 5         OHQ430 Do you have a lower removable partial or full dent
#> 6         OHQ440             Do you usually wear it during the day?
#> 7       OHX01HTC            Tooth Count: Upper right 3rd molar (3M)
#> 8       OHX02HTC            Tooth Count: Upper right 2nd molar (2M)
#> 9       OHX03HTC            Tooth Count: Upper right 1st molar (1M)
#> 10      OHX04HTC         Tooth Count: Upper right 2nd bicuspid (2B)
#> 11      OHX05HTC         Tooth Count: Upper right 1st bicuspid (1B)
#> 12      OHX06HTC                Tooth Count: Upper right cuspid (C)
#> 13      OHX07HTC      Tooth Count: Upper right lateral incisor (LI)
#> 14      OHX08HTC      Tooth Count: Upper right central incisor (CI)
#> 15      OHX09HTC       Tooth Count: Upper left central incisor (CI)
#> 16      OHX10HTC       Tooth Count: Upper left lateral incisor (LI)
#> 17      OHX11HTC                 Tooth Count: Upper left cuspid (C)
#> 18      OHX12HTC          Tooth Count: Upper left 1st bicuspid (1B)
#> 19      OHX13HTC          Tooth Count: Upper left 2nd bicuspid (2B)
#> 20      OHX14HTC             Tooth Count: Upper left 1st molar (1M)
#> 21      OHX15HTC             Tooth Count: Upper left 2nd molar (2M)
#> 22      OHX16HTC             Tooth Count: Upper left 3rd molar (3M)
#> 23      OHX17HTC             Tooth Count: Lower left 3rd molar (3M)
#> 24      OHX18HTC             Tooth Count: Lower left 2nd molar (2M)
#> 25      OHX19HTC             Tooth Count: Lower left 1st molar (1M)
#> 26      OHX20HTC          Tooth Count: Lower left 2nd bicuspid (2B)
#> 27      OHX21HTC          Tooth Count: Lower left 1st bicuspid (1B)
#> 28      OHX22HTC                 Tooth Count: Lower left cuspid (C)
#> 29      OHX23HTC       Tooth Count: Lower left lateral incisor (LI)
#> 30      OHX24HTC       Tooth Count: Lower left central incisor (CI)
#> 31      OHX25HTC      Tooth Count: Lower right central incisor (CI)
#> 32      OHX26HTC      Tooth Count: Lower right lateral incisor (LI)
#> 33      OHX27HTC                Tooth Count: Lower right cuspid (C)
#> 34      OHX28HTC         Tooth Count: Lower right 1st bicuspid (1B)
#> 35      OHX29HTC         Tooth Count: Lower right 2nd bicuspid (2B)
#> 36      OHX30HTC            Tooth Count: Lower right 1st molar (1M)
#> 37      OHX31HTC            Tooth Count: Lower right 2nd molar (2M)
#> 38      OHX32HTC            Tooth Count: Lower right 3rd molar (3M)
#> 39      OHXDECAY   Presence of at least one tooth with dental decay
#> 40      OHXFCANT                                     Anterior Zone:
#> 41      OHXFCLZ1                                      Left - Zone 1
#> 42      OHXFCLZ2                                      Left - Zone 2
#> 43      OHXFCLZ3                                      Left - Zone 3
#> 44      OHXFCLZ4                                      Left - Zone 4
#> 45      OHXFCLZ5                                      Left - Zone 5
#> 46      OHXFCLZ6                                      Left - Zone 6
#> 47      OHXFCLZ7                                      Left - Zone 7
#> 48      OHXFCLZ8                                      Left - Zone 8
#> 49      OHXFCRZ1                                     Right - Zone 1
#> 50      OHXFCRZ2                                     Right - Zone 2
#> 51      OHXFCRZ3                                     Right - Zone 3
#> 52      OHXFCRZ4                                     Right - Zone 4
#> 53      OHXFCRZ5                                     Right - Zone 5
#> 54      OHXFCRZ6                                     Right - Zone 6
#> 55      OHXFCRZ7                                     Right - Zone 7
#> 56      OHXFCRZ8                                     Right - Zone 8
#> 57       OHXREST Presence of at least one tooth with a dental resto
#> 58       OHXSEAL Presence of at least one tooth with a dental preve
#> 59          SEQN                        Respondent sequence number.
#>    Data.File.Name Data.File.Description Begin.Year EndYear   Component
#> 1           OHX_E           Oral Health       2007    2008 Examination
#> 2           OHX_E           Oral Health       2007    2008 Examination
#> 3           OHX_E           Oral Health       2007    2008 Examination
#> 4           OHX_E           Oral Health       2007    2008 Examination
#> 5           OHX_E           Oral Health       2007    2008 Examination
#> 6           OHX_E           Oral Health       2007    2008 Examination
#> 7           OHX_E           Oral Health       2007    2008 Examination
#> 8           OHX_E           Oral Health       2007    2008 Examination
#> 9           OHX_E           Oral Health       2007    2008 Examination
#> 10          OHX_E           Oral Health       2007    2008 Examination
#> 11          OHX_E           Oral Health       2007    2008 Examination
#> 12          OHX_E           Oral Health       2007    2008 Examination
#> 13          OHX_E           Oral Health       2007    2008 Examination
#> 14          OHX_E           Oral Health       2007    2008 Examination
#> 15          OHX_E           Oral Health       2007    2008 Examination
#> 16          OHX_E           Oral Health       2007    2008 Examination
#> 17          OHX_E           Oral Health       2007    2008 Examination
#> 18          OHX_E           Oral Health       2007    2008 Examination
#> 19          OHX_E           Oral Health       2007    2008 Examination
#> 20          OHX_E           Oral Health       2007    2008 Examination
#> 21          OHX_E           Oral Health       2007    2008 Examination
#> 22          OHX_E           Oral Health       2007    2008 Examination
#> 23          OHX_E           Oral Health       2007    2008 Examination
#> 24          OHX_E           Oral Health       2007    2008 Examination
#> 25          OHX_E           Oral Health       2007    2008 Examination
#> 26          OHX_E           Oral Health       2007    2008 Examination
#> 27          OHX_E           Oral Health       2007    2008 Examination
#> 28          OHX_E           Oral Health       2007    2008 Examination
#> 29          OHX_E           Oral Health       2007    2008 Examination
#> 30          OHX_E           Oral Health       2007    2008 Examination
#> 31          OHX_E           Oral Health       2007    2008 Examination
#> 32          OHX_E           Oral Health       2007    2008 Examination
#> 33          OHX_E           Oral Health       2007    2008 Examination
#> 34          OHX_E           Oral Health       2007    2008 Examination
#> 35          OHX_E           Oral Health       2007    2008 Examination
#> 36          OHX_E           Oral Health       2007    2008 Examination
#> 37          OHX_E           Oral Health       2007    2008 Examination
#> 38          OHX_E           Oral Health       2007    2008 Examination
#> 39          OHX_E           Oral Health       2007    2008 Examination
#> 40          OHX_E           Oral Health       2007    2008 Examination
#> 41          OHX_E           Oral Health       2007    2008 Examination
#> 42          OHX_E           Oral Health       2007    2008 Examination
#> 43          OHX_E           Oral Health       2007    2008 Examination
#> 44          OHX_E           Oral Health       2007    2008 Examination
#> 45          OHX_E           Oral Health       2007    2008 Examination
#> 46          OHX_E           Oral Health       2007    2008 Examination
#> 47          OHX_E           Oral Health       2007    2008 Examination
#> 48          OHX_E           Oral Health       2007    2008 Examination
#> 49          OHX_E           Oral Health       2007    2008 Examination
#> 50          OHX_E           Oral Health       2007    2008 Examination
#> 51          OHX_E           Oral Health       2007    2008 Examination
#> 52          OHX_E           Oral Health       2007    2008 Examination
#> 53          OHX_E           Oral Health       2007    2008 Examination
#> 54          OHX_E           Oral Health       2007    2008 Examination
#> 55          OHX_E           Oral Health       2007    2008 Examination
#> 56          OHX_E           Oral Health       2007    2008 Examination
#> 57          OHX_E           Oral Health       2007    2008 Examination
#> 58          OHX_E           Oral Health       2007    2008 Examination
#> 59          OHX_E           Oral Health       2007    2008 Examination
#>    Use.Constraints
#> 1             None
#> 2             None
#> 3             None
#> 4             None
#> 5             None
#> 6             None
#> 7             None
#> 8             None
#> 9             None
#> 10            None
#> 11            None
#> 12            None
#> 13            None
#> 14            None
#> 15            None
#> 16            None
#> 17            None
#> 18            None
#> 19            None
#> 20            None
#> 21            None
#> 22            None
#> 23            None
#> 24            None
#> 25            None
#> 26            None
#> 27            None
#> 28            None
#> 29            None
#> 30            None
#> 31            None
#> 32            None
#> 33            None
#> 34            None
#> 35            None
#> 36            None
#> 37            None
#> 38            None
#> 39            None
#> 40            None
#> 41            None
#> 42            None
#> 43            None
#> 44            None
#> 45            None
#> 46            None
#> 47            None
#> 48            None
#> 49            None
#> 50            None
#> 51            None
#> 52            None
#> 53            None
#> 54            None
#> 55            None
#> 56            None
#> 57            None
#> 58            None
#> 59            None
nhanesTableVars('DEMO', 'DEMO_F', namesonly = TRUE)
#>  [1] "AIALANG"  "DMDBORN2" "DMDCITZN" "DMDEDUC2" "DMDEDUC3" "DMDFMSIZ"
#>  [7] "DMDHHSIZ" "DMDHRAGE" "DMDHRBR2" "DMDHREDU" "DMDHRGND" "DMDHRMAR"
#> [13] "DMDHSEDU" "DMDMARTL" "DMDSCHOL" "DMDYRSUS" "DMQMILIT" "FIAINTRP"
#> [19] "FIALANG"  "FIAPROXY" "INDFMIN2" "INDFMPIR" "INDHHIN2" "MIAINTRP"
#> [25] "MIALANG"  "MIAPROXY" "RIAGENDR" "RIDAGEEX" "RIDAGEMN" "RIDAGEYR"
#> [31] "RIDEXMON" "RIDEXPRG" "RIDRETH1" "RIDSTATR" "SDDSRVYR" "SDMVPSU" 
#> [37] "SDMVSTRA" "SEQN"     "SIAINTRP" "SIALANG"  "SIAPROXY" "WTINT2YR"
#> [43] "WTMEC2YR"