Returns a list of table names for the specified survey group.

Enables quick display of all available tables in the survey group.

nhanesTables(
  data_group,
  year,
  nchar = 100,
  details = FALSE,
  namesonly = FALSE,
  includerdc = FALSE
)

Arguments

data_group
The type of survey (DEMOGRAPHICS, DIETARY, EXAMINATION, LABORATORY, QUESTIONNAIRE). Abbreviated terms may also be used: (DEMO, DIET, EXAM, LAB, Q).

year
The year in yyyy format where 1999 <= yyyy <= 2014.

nchar
Truncates the table description to a max length of nchar.

details
If TRUE then a more detailed description of the tables is returned (default=FALSE).

namesonly
If TRUE then only the table names are returned (default=FALSE).

includerdc
If TRUE then RDC only tables are included in list (default=FALSE).

Value

The names of the tables in the specified survey group.

Details

Data are retrieved via web scraping using html wrappers from package rvest. It is often useful to display the table names in an NHANES survey. In effect this is a convenient way to browse the available NHANES tables.

Examples

nhanesTables('EXAM', 2007)
#>    Data.File.Name                         Data.File.Description
#> 1           VIX_E                                        Vision
#> 2           BPX_E                                Blood Pressure
#> 3        OPXFDT_E Ophthalmology - Frequency Doubling Technology
#> 4           OHX_E                                   Oral Health
#> 5        DXXFEM_E      Dual Energy X-ray Absorptiometry - Femur
#> 6           BMX_E                                 Body Measures
#> 7           AUX_E                                    Audiometry
#> 8        AUXTYM_E                     Audiometry - Tympanometry
#> 9           ENX_E                          Exhaled Nitric Oxide
#> 10          SPX_E      Spirometry - Pre and Post-Bronchodilator
#> 11       SPXRAW_E                   Spirometry - Raw Curve Data
#> 12        AUXAR_E                  Audiometry - Acoustic Reflex
#> 13       OPXRET_E               Ophthalmology - Retinal Imaging
#> 14       DXXSPN_E      Dual Energy X-ray Absorptiometry - Spine
nhanesTables('LAB', 2009, details=TRUE, includerdc=TRUE)
#>    Data.File.Name
#> 1         B27_F_R
#> 2        UCPREG_F
#> 3           HDL_F
#> 4        FOLATE_F
#> 5          HEPA_F
#> 6         HEPBD_F
#> 7           HIV_F
#> 8           HSV_F
#> 9          MMRV_F
#> 10         PbCd_F
#> 11        TCHOL_F
#> 12        TGEMA_F
#> 13          UHG_F
#> 14          CBC_F
#> 15       COTNAL_F
#> 16          GHB_F
#> 17        ORHPV_F
#> 18          IHg_F
#> 19       BIOPRO_F
#> 20          UAS_F
#> 21          UHM_F
#> 22          PSA_F
#> 23       ALB_CR_F
#> 24       CHLMDA_F
#> 25          CRP_F
#> 26       FASTQX_F
#> 27          GLU_F
#> 28       HEPB_S_F
#> 29         HEPC_F
#> 30          UIO_F
#> 31       UCFLOW_F
#> 32       UCOSMO_F
#> 33         OGTT_F
#> 34       TRIGLY_F
#> 35       THYROD_F
#> 36        HSV_F_R
#> 37       CHLM_F_R
#> 38       OHPV_F_R
#> 39          EPH_F
#> 40           PP_F
#> 41          PFC_F
#> 42       HPVSWR_F
#> 43       FERTIN_F
#> 44          TFR_F
#> 45       HPVS_F_R
#> 46         HEPE_F
#> 47         ApoB_F
#> 48        SSEBV_F
#> 49       PHTHTE_F
#> 50        PHYTO_F
#> 51        SSHPV_F
#> 52         CAFE_F
#> 53       SSTOXO_F
#> 54       HPVSER_F
#> 55        SER_F_R
#> 56       SSHP_F_R
#> 57        PERNT_F
#> 58       SSPOLI_F
#> 59       SSIF_F_R
#> 60       UPHOPM_F
#> 61          TFA_F
#> 62         DEET_F
#> 63          PAH_F
#> 64       VIT_B6_F
#> 65          VID_F
#> 66          VOC_F
#> 67       VOCMWB_F
#> 68       DOXPOL_F
#> 69        VOCWB_F
#> 70       PSTPOL_F
#> 71       BFRPOL_F
#> 72       PCBPOL_F
#> 73       POOLTF_F
#> 74         SSKL_F
#>                                                                                   Data.File.Description
#> 1                                                                          Human Leukocyte Antigen B-27
#> 2                                                                                Pregnancy Test - Urine
#> 3                                                                                     Cholesterol - HDL
#> 4                                                                                  Folate - RBC & Serum
#> 5                                                                                  Hepatitis A Antibody
#> 6                                     Hepatitis B: Core Antibody, Surface Antigen; Hepatitis D Antibody
#> 7                                                                                     HIV Antibody Test
#> 8                                                                  Herpes Simplex Virus Type-1 & Type-2
#> 9                                                                   Measles, Mumps, Rubella & Varicella
#> 10                                                               Cadmium, Lead, & Total Mercury - Blood
#> 11                                                                                  Cholesterol - Total
#> 12                    Tissue Transglutaminase Assay (IgA-TTG) & IgA Endomyseal Antibody Assay (IgA EMA)
#> 13                                                                           Mercury - Inorganic, Urine
#> 14                                          Complete Blood Count with 5-part Differential - Whole Blood
#> 15                                                                Cotinine - Serum & Total NNAL - Urine
#> 16                                                                                      Glycohemoglobin
#> 17                                                              Human Papillomavirus (HPV) - Oral Rinse
#> 18                                                                          Mercury - Inorganic - Blood
#> 19                                                                        Standard Biochemistry Profile
#> 20                                                                 Arsenics - Total & Speciated - Urine
#> 21                                                                                       Metals - Urine
#> 22                                                                      Prostate Specific Antigen (PSA)
#> 23                                                                         Albumin & Creatinine - Urine
#> 24                                                                                    Chlamydia - Urine
#> 25                                                                             C-Reactive Protein (CRP)
#> 26                                                                                Fasting Questionnaire
#> 27                                                                     Plasma Fasting Glucose & Insulin
#> 28                                                                         Hepatitis B Surface Antibody
#> 29                                           Hepatitis C: Confirmed Antibody, RNA (HCV-RNA), & Genotype
#> 30                                                                                       Iodine - Urine
#> 31                                                                                      Urine Flow Rate
#> 32                                                                                   Osmolality - Urine
#> 33                                                                          Oral Glucose Tolerance Test
#> 34                                                                    Cholesterol - LDL & Triglycerides
#> 35                                                                                      Thyroid Profile
#> 36                                                                  Herpes Simplex Virus Type-2 - Youth
#> 37                                                                Chlamydia & Gonorrhea - Urine - Youth
#> 38                                                      Human Papillomavirus (HPV) - Oral Rinse - Youth
#> 39                                                                                Environmental Phenols
#> 40                                                                   Pesticides - Environmental - Urine
#> 41                                                                            Polyfluoroalkyl Chemicals
#> 42            Human Papillomavirus (HPV) DNA - Vaginal Swab: Digene Hybrid Capture & Roche Linear Array
#> 43                                                                                             Ferritin
#> 44                                                                                 Transferrin Receptor
#> 45    Human Papillomavirus (HPV) DNA - Vaginal Swab: Digene Hybrid Capture & Roche Linear Array - Youth
#> 46                                                                   Hepatitis E : IgG & IgM Antibodies
#> 47                                                                                     Apolipoprotein B
#> 48                                                       Epstein-Barr Virus (VCA IgG) - Serum (Surplus)
#> 49                                                                                   Phthalates - Urine
#> 50                                                                               Phytoestrogens - Urine
#> 51         Human Papillomavirus (HPV) Viral Load - Oral Rinse - Oral High-Risk HPV Infections (Surplus)
#> 52                                                              Caffeine & Caffeine Metabolites - Urine
#> 53                                                         Toxoplasma Gondii Antibody - Serum (Surplus)
#> 54                                         Human Papillomavirus (HPV) - 6, 11, 16 & 18 Antibody - Serum
#> 55                                 Human Papillomavirus (HPV) - 6, 11, 16 & 18 Antibody - Serum - Youth
#> 56 Human Papillomavirus (HPV) Viral Load - Oral Rinse - Oral High-Risk HPV Infections - Youth (Surplus)
#> 57                                                           Perchlorate, Nitrate & Thiocyanate - Urine
#> 58                                          Poliovirus Serotypes 1, 2, & 3 Antibodies - Serum (Surplus)
#> 59   Influenza A (H3N2) Variant Virus & Historical Seasonal H3N2 Influenza Viruses Antibodies (Surplus)
#> 60                                                    Pyrethroids, Herbicides, & OP Metabolites - Urine
#> 61                                                                                    Trans Fatty Acids
#> 62                                                                                 DEET and Metabolites
#> 63                                                                     Polyaromatic Hydrocarbons -Urine
#> 64                                                                                           Vitamin B6
#> 65                                                                                            Vitamin D
#> 66                                                            Volatile Organic Compounds (VOCs) - Water
#> 67                        Volatile Organic Compounds (VOCs) - Trihalomethanes/MTBE/Nitromethane - Blood
#> 68 Polychlorinated dibenzo-p-dioxins (PCDDs), Dibenzofurans (PCDFs) & Coplanar Polychlorinated Biphenyl
#> 69                                                            Volatile Organic Compounds (VOCs) - Blood
#> 70                                              Pesticides - Organochlorine Pesticides - Pooled Samples
#> 71                                                  Brominated Flame Retardants (BFRs) - Pooled Samples
#> 72 Non-dioxin-like Polychlorinated Biphenyls & Mono-ortho-substituted Polychlorinated Biphenyls - Poole
#> 73                                                                 Pooled-Sample Technical Support File
#> 74                                                                             Klotho - Serum (Surplus)
#>    Begin.Year EndYear  Component Use.Constraints
#> 1        2009    2009 Laboratory        RDC Only
#> 2        2009    2010 Laboratory            None
#> 3        2009    2010 Laboratory            None
#> 4        2009    2010 Laboratory            None
#> 5        2009    2010 Laboratory            None
#> 6        2009    2010 Laboratory            None
#> 7        2009    2010 Laboratory            None
#> 8        2009    2010 Laboratory            None
#> 9        2009    2010 Laboratory            None
#> 10       2009    2010 Laboratory            None
#> 11       2009    2010 Laboratory            None
#> 12       2009    2010 Laboratory            None
#> 13       2009    2010 Laboratory            None
#> 14       2009    2010 Laboratory            None
#> 15       2009    2010 Laboratory            None
#> 16       2009    2010 Laboratory            None
#> 17       2009    2010 Laboratory            None
#> 18       2009    2010 Laboratory            None
#> 19       2009    2010 Laboratory            None
#> 20       2009    2010 Laboratory            None
#> 21       2009    2010 Laboratory            None
#> 22       2009    2010 Laboratory            None
#> 23       2009    2010 Laboratory            None
#> 24       2009    2010 Laboratory            None
#> 25       2009    2010 Laboratory            None
#> 26       2009    2010 Laboratory            None
#> 27       2009    2010 Laboratory            None
#> 28       2009    2010 Laboratory            None
#> 29       2009    2010 Laboratory            None
#> 30       2009    2010 Laboratory            None
#> 31       2009    2010 Laboratory            None
#> 32       2009    2010 Laboratory            None
#> 33       2009    2010 Laboratory            None
#> 34       2009    2010 Laboratory            None
#> 35       2009    2010 Laboratory            None
#> 36       2009    2010 Laboratory        RDC Only
#> 37       2009    2010 Laboratory        RDC Only
#> 38       2009    2010 Laboratory        RDC Only
#> 39       2009    2010 Laboratory            None
#> 40       2009    2010 Laboratory            None
#> 41       2009    2010 Laboratory            None
#> 42       2009    2010 Laboratory            None
#> 43       2009    2010 Laboratory            None
#> 44       2009    2010 Laboratory            None
#> 45       2009    2010 Laboratory        RDC Only
#> 46       2009    2010 Laboratory            None
#> 47       2009    2010 Laboratory            None
#> 48       2009    2010 Laboratory            None
#> 49       2009    2010 Laboratory            None
#> 50       2009    2010 Laboratory            None
#> 51       2009    2010 Laboratory            None
#> 52       2009    2010 Laboratory            None
#> 53       2009    2010 Laboratory            None
#> 54       2009    2010 Laboratory            None
#> 55       2009    2010 Laboratory        RDC Only
#> 56       2009    2010 Laboratory        RDC Only
#> 57       2009    2010 Laboratory            None
#> 58       2009    2010 Laboratory            None
#> 59       2010    2010 Laboratory        RDC Only
#> 60       2009    2010 Laboratory            None
#> 61       2009    2010 Laboratory            None
#> 62       2009    2010 Laboratory            None
#> 63       2009    2010 Laboratory            None
#> 64       2009    2010 Laboratory            None
#> 65       2009    2010 Laboratory            None
#> 66       2009    2010 Laboratory            None
#> 67       2009    2010 Laboratory            None
#> 68       2009    2010 Laboratory            None
#> 69       2009    2010 Laboratory            None
#> 70       2009    2010 Laboratory            None
#> 71       2009    2010 Laboratory            None
#> 72       2009    2010 Laboratory            None
#> 73       2009    2010 Laboratory            None
#> 74       2009    2010 Laboratory            None
nhanesTables('Q', 2005, namesonly=TRUE)
#>  [1] "ACQ_D"    "ALQ_D"    "AGQ_D"    "BHQ_D"    "BPQ_D"    "CDQ_D"   
#>  [7] "HSQ_D"    "DPQ_D"    "DEQ_D"    "DBQ_D"    "AUQ_D"    "DIQ_D"   
#> [13] "DUQ_D"    "ECQ_D"    "FSQ_D"    "HIQ_D"    "HCQ_D"    "HUQ_D"   
#> [19] "HOQ_D"    "IMQ_D"    "KIQ_U_D"  "MCQ_D"    "OHQ_D"    "OSQ_D"   
#> [25] "PUQMEC_D" "PAQ_D"    "PAQIAF_D" "PFQ_D"    "RXQ_RX_D" "KIQ_P_D" 
#> [31] "PSQ_D"    "RHQ_D"    "RDQ_D"    "SXQ_D"    "SLQ_D"    "SMQ_D"   
#> [37] "SMQFAM_D" "SMQRTU_D" "SSQ_D"    "VIQ_D"    "WHQMEC_D" "WHQ_D"   
#> [43] "OCQ_D"    "RXQ_DRUG"